Adjusted implicit casting of par values (bug in BMacLayer.cc where the "= hasPar() ? par() : int" construct was used and the compiler does a implicit int cast instead of double). Adjusted also the .ned settings for integer values which was defined as double values. (see https://sourceforge.net/tracker/index.php?func=detail&aid=3496931&group_id=176959&atid=879320)

+++ src/modules/mac/csma.cc
@@ -66,7 +66,7 @@ void csma::initialize(int stage) {
 		macMaxCSMABackoffs = par("macMaxCSMABackoffs");
 		macMaxFrameRetries = par("macMaxFrameRetries");
 		macAckWaitDuration = par("macAckWaitDuration").doubleValue();
-		aUnitBackoffPeriod = par("aUnitBackoffPeriod");
+		aUnitBackoffPeriod = par("aUnitBackoffPeriod").doubleValue();
 		ccaDetectionTime = par("ccaDetectionTime").doubleValue();
 		rxSetupTime = par("rxSetupTime").doubleValue();
 		aTurnaroundTime = par("aTurnaroundTime").doubleValue();
